ccompiler stringbuffer tests

+++ core/ccompiler/include/ccompiler/stringbuffer.h
@@ -2,11 +2,13 @@
 #ifndef STRINGBUFFER_H
 #define STRINGBUFFER_H
 
+#include "inttypes.h"
+
 typedef struct StringBuffer
 {
     char *buf;
-    int length;
-    int capacity;
+    size_t length;
+    size_t capacity;
 } stringbuffer_t;
 
 
@@ -15,7 +17,7 @@ void stringbuffer_free (stringbuffer_t *stringbuffer);
 
 void stringbuffer_appendf (stringbuffer_t *stringbuffer, const char *fmt, ...);
 void stringbuffer_append (stringbuffer_t *stringbuffer, const char *str);
-void stringbuffer_appendl (stringbuffer_t *stringbuffer, const char *str, const int len);
+void stringbuffer_appendl (stringbuffer_t *stringbuffer, const char *str, const size_t len);
 
 void stringbuffer_clear (stringbuffer_t *stringbuffer);
